+++ src/gql/resolvers/elements/User.php
@@ -7,6 +7,7 @@
 
 namespace craft\gql\resolvers\elements;
 
+use Craft;
 use craft\db\Table;
 use craft\elements\User as UserElement;
 use craft\gql\base\ElementResolver;
@@ -52,7 +53,12 @@ class User extends ElementResolver
             }
 
             $pairs = GqlHelper::extractAllowedEntitiesFromSchema('read');
-            $allowedGroupIds = array_values(Db::idsByUids(Table::USERGROUPS, $pairs['usergroups']));
+
+            $userGroupsService = Craft::$app->getUserGroups();
+            $allowedGroupIds = array_filter(array_map(function(string $uid) use ($userGroupsService) {
+                $userGroupsService = $userGroupsService->getGroupByUid($uid);
+                return $userGroupsService->id ?? null;
+            }, $pairs['usergroups']));
 
             $query->groupId = $query->groupId ? array_intersect($allowedGroupIds, (array)$query->groupId) : $allowedGroupIds;
         }